\name{clg}
\alias{clg}
\alias{PROBj2k}
\title{Probability to lapse}
\description{
		Cumulative logistic function
	}
\usage{
PROBj2k(x, j, k, param_j)
}
\arguments{
  \item{x}{vector \code{x}.}
  \item{j}{integer \code{j}.}
  \item{k}{integer \code{k}.}
  \item{param_j}{vector \code{param_j}.}    
}
\value{
A numeric.
}
\author{
 Christophe Dutang
}
\keyword{math}


